Backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
pydantic==2.9.2

